requirements.txt
azure-cognitiveservices-speech
python-dotenv
urllib3